requirements.txt
Django==1.9b1
MySQL-python==1.2.5
django-debug-toolbar==1.4
numpy==1.10.1
biopython==1.66
reportlab==3.2.0
matplotlib==1.5.0
django-bootstrap-form==3.2
django-formtools==1.0
uniprot_tools=0.4.1
